@@ -1,8 +1,6 @@
 module github.com/badtuxx/girus-cli
 
-go 1.23.0
-
-toolchain go1.23.7
+go 1.24
 
 require (
 	github.com/go-git/go-git/v5 v5.15.0
